chatbot_agendamento/requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.70.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
ics==0.7.2
